+++ src/features/picomatch.d.ts
@@ -1,0 +1,1 @@
+declare module 'picomatch/lib/picomatch.js'