+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Change-Log:
 
+### Version 0.0.13 (2024-06-24)
+
+- Update readPDF.ts to handle PDF content display
+- Add picomatch type definition file
+- Refined the prefix prompt
+
+
 ### Version 0.0.12 (2024-06-19)
 
 - Add logging to readPDF and webSearch features